requirements.txt
langchain
langchain-google-genai
langchain-community
python-dotenv
SQLAlchemy
pypdf
gradio
